+++ gateway/graphql/paginated_resolvers.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
+	"time"
 
 	"code.vegaprotocol.io/data-node/logging"
 	"code.vegaprotocol.io/data-node/vegatime"
@@ -254,6 +255,11 @@ func (r *myPaginatedMarketResolver) RiskFactors(ctx context.Context, obj *types.
 	}
 
 	return rf.RiskFactor, nil
+}
+
+func (r *myPaginatedMarketResolver) CandlesConnection(ctx context.Context, market *types.Market, sinceRaw string, toRaw *string,
+	interval Interval, pagination *v2.Pagination) (*v2.CandleDataConnection, error) {
+	return handleCandleConnectionRequest(ctx, r.tradingDataClientV2, market, sinceRaw, toRaw, interval, pagination)
 }
 
 type myPaginatedPartyResolver VegaResolverRoot
@@ -756,3 +762,63 @@ func (r *myPaginatedOrderResolver) LiquidityProvision(ctx context.Context, obj *
 }
 
 // END: Paginated Order Resolver
+
+func handleCandleConnectionRequest(ctx context.Context, client TradingDataServiceClientV2, market *types.Market, sinceRaw string, toRaw *string,
+	interval Interval, pagination *v2.Pagination) (*v2.CandleDataConnection, error) {
+	pInterval, err := convertIntervalToProto(interval)
+	if err != nil {
+		return nil, fmt.Errorf("could not convert interval: %w", err)
+	}
+
+	since, err := vegatime.Parse(sinceRaw)
+	if err != nil {
+		return nil, err
+	}
+
+	to := time.Unix(0, 0)
+	if toRaw != nil {
+		to, err = vegatime.Parse(*toRaw)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	var mkt string
+	if market != nil {
+		mkt = market.Id
+	}
+
+	candlesForMktReq := v2.GetCandlesForMarketRequest{MarketId: mkt}
+	candlesForMktResp, err := client.GetCandlesForMarket(ctx, &candlesForMktReq)
+
+	if err != nil {
+		return nil, fmt.Errorf("could not retrieve candles for market %s: %w", mkt, err)
+	}
+
+	candleID := ""
+
+	for _, c4m := range candlesForMktResp.IntervalToCandleId {
+		if c4m.Interval == string(interval) {
+			candleID = c4m.CandleId
+			break
+		}
+	}
+
+	if candleID == "" {
+		return nil, fmt.Errorf("could not find candle for market %s and interval %s", mkt, interval)
+	}
+
+	req := v2.GetCandleDataRequest{
+		CandleId:      candleID,
+		FromTimestamp: since.Unix(),
+		ToTimestamp:   to.Unix(),
+		Interval:      pInterval,
+		Pagination:    pagination,
+	}
+	resp, err := client.GetCandleData(ctx, &req)
+	if err != nil {
+		return nil, fmt.Errorf("could not retrieve candles for market %s: %w", mkt, err)
+	}
+
+	return resp.Candles, nil
+}