+++ sqlstore/candles_test.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"code.vegaprotocol.io/data-node/candlesv2"
+	"github.com/stretchr/testify/require"
 
 	types "code.vegaprotocol.io/protos/vega"
 
@@ -61,12 +62,13 @@ func TestCandlesPagination(t *testing.T) {
 	insertCandlesTestData(t, tradeStore, startTime, totalBlocks, tradesPerBlock, startPrice, priceIncrement, size, blockIntervalDur)
 
 	_, candleId, _ := candleStore.GetCandleIdForIntervalAndMarket(context.Background(), "1 Minute", testMarket)
-	candles, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, nil,
-		nil, entities.OffsetPagination{
-			Skip:       0,
-			Limit:      10,
-			Descending: true,
-		})
+	first := int32(10)
+	pagination, err := entities.NewCursorPagination(&first, nil, nil, nil)
+	require.NoError(t, err)
+
+	candles, _, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, nil,
+		nil, pagination)
+
 	if err != nil {
 		t.Fatalf("failed to get candles with pagination:%s", err)
 	}
@@ -74,12 +76,13 @@ func TestCandlesPagination(t *testing.T) {
 	assert.Equal(t, 10, len(candles))
 	lastCandle := candles[9]
 
-	candles, err = candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, nil,
-		nil, entities.OffsetPagination{
-			Skip:       9,
-			Limit:      5,
-			Descending: true,
-		})
+	first = int32(5)
+	after := entities.NewCursor(candles[8].PeriodStart.Format(time.RFC3339Nano)).Encode()
+
+	pagination, _ = entities.NewCursorPagination(&first, &after, nil, nil)
+
+	candles, _, err = candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, nil,
+		nil, pagination)
 
 	if err != nil {
 		t.Fatalf("failed to get candles with pagination:%s", err)
@@ -112,8 +115,9 @@ func TestCandlesGetForEmptyInterval(t *testing.T) {
 		t.Fatalf("getting existing candleDescriptor id:%s", err)
 	}
 
-	candles, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime,
-		nil, entities.OffsetPagination{})
+	pagination, _ := entities.NewCursorPagination(nil, nil, nil, nil)
+	candles, _, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime,
+		nil, pagination)
 	if err != nil {
 		t.Fatalf("failed to get candles:%s", err)
 	}
@@ -139,13 +143,11 @@ func TestCandlesGetLatest(t *testing.T) {
 	insertCandlesTestData(t, tradeStore, startTime, 90, 3, startPrice, priceIncrement, size,
 		1*time.Second)
 
+	last := int32(1)
+	pagination, _ := entities.NewCursorPagination(nil, nil, &last, nil)
 	_, candleId, _ := candleStore.GetCandleIdForIntervalAndMarket(context.Background(), "1 Minute", testMarket)
-	candles, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime,
-		nil, entities.OffsetPagination{
-			Skip:       0,
-			Limit:      1,
-			Descending: true,
-		})
+	candles, _, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime,
+		nil, pagination)
 	if err != nil {
 		t.Fatalf("failed to get candles:%s", err)
 	}
@@ -190,9 +192,11 @@ func testInterval(t *testing.T, tradeDataStartTime time.Time, fromTime *time.Tim
 ) {
 	intervalDur := time.Duration(intervalSeconds) * time.Second
 
+	pagination, _ := entities.NewCursorPagination(nil, nil, nil, nil)
+	//entities.OffsetPagination{}
 	_, candleId, _ := candleStore.GetCandleIdForIntervalAndMarket(context.Background(), interval, testMarket)
-	candles, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, fromTime,
-		toTime, entities.OffsetPagination{})
+	candles, _, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, fromTime,
+		toTime, pagination)
 	if err != nil {
 		t.Fatalf("failed to get candles:%s", err)
 	}
@@ -313,4 +317,92 @@ func createTestTrade(t *testing.T, price int, size int, block entities.Block, se
 		t.Fatalf("failed to create trade from proto:%s", err)
 	}
 	return trade
+}
+
+func TestCandlesCursorPagination(t *testing.T) {
+	defer DeleteEverything()
+
+	candleStore := sqlstore.NewCandles(context.Background(), connectionSource, candlesv2.NewDefaultConfig().CandleStore)
+	tradeStore := sqlstore.NewTrades(connectionSource)
+
+	startTime := time.Unix(StartTime, 0)
+	insertCandlesTestData(t, tradeStore, startTime, totalBlocks, tradesPerBlock, startPrice, priceIncrement, size, blockIntervalDur)
+
+	_, candleId, err := candleStore.GetCandleIdForIntervalAndMarket(context.Background(), "1 Minute", testMarket)
+	if err != nil {
+		t.Fatalf("getting existing candleDescriptor id:%s", err)
+	}
+
+	pagination, _ := entities.NewCursorPagination(nil, nil, nil, nil)
+	// retrieve all candles without pagination to use for test validation
+	allCandles, _, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime,
+		nil, pagination)
+	if err != nil {
+		t.Fatalf("failed to get candles:%s", err)
+	}
+
+	require.Equal(t, 167, len(allCandles))
+
+	t.Run("should return the first candles when first is provided with no after", func(t *testing.T) {
+		first := int32(10)
+		pagination, err := entities.NewCursorPagination(&first, nil, nil, nil)
+		require.NoError(t, err)
+		candles, pageInfo, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime, nil, pagination)
+		require.NoError(t, err)
+		require.Equal(t, 10, len(candles))
+		assert.Equal(t, allCandles[0], candles[0])
+		assert.Equal(t, allCandles[9], candles[9])
+		assert.Equal(t, pageInfo.StartCursor, entities.NewCursor(allCandles[0].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.Equal(t, pageInfo.EndCursor, entities.NewCursor(allCandles[9].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.False(t, pageInfo.HasPreviousPage)
+		assert.True(t, pageInfo.HasNextPage)
+	})
+
+	t.Run("should return the last page of candles when last is provided with no before", func(t *testing.T) {
+		last := int32(10)
+		pagination, err := entities.NewCursorPagination(nil, nil, &last, nil)
+		require.NoError(t, err)
+		candles, pageInfo, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime, nil, pagination)
+		require.NoError(t, err)
+		require.Equal(t, 10, len(candles))
+		assert.Equal(t, allCandles[157], candles[0])
+		assert.Equal(t, allCandles[166], candles[9])
+		assert.Equal(t, pageInfo.StartCursor, entities.NewCursor(allCandles[157].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.Equal(t, pageInfo.EndCursor, entities.NewCursor(allCandles[166].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.True(t, pageInfo.HasPreviousPage)
+		assert.False(t, pageInfo.HasNextPage)
+	})
+
+	t.Run("should return the requested page of candles when first and after are provided", func(t *testing.T) {
+		first := int32(10)
+		after := entities.NewCursor(allCandles[99].PeriodStart.Format(time.RFC3339Nano)).Encode()
+		pagination, err := entities.NewCursorPagination(&first, &after, nil, nil)
+		require.NoError(t, err)
+		candles, pageInfo, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime, nil, pagination)
+		require.NoError(t, err)
+		require.Equal(t, 10, len(candles))
+		assert.Equal(t, allCandles[100], candles[0])
+		assert.Equal(t, allCandles[109], candles[9])
+		assert.Equal(t, pageInfo.StartCursor, entities.NewCursor(allCandles[100].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.Equal(t, pageInfo.EndCursor, entities.NewCursor(allCandles[109].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.True(t, pageInfo.HasPreviousPage)
+		assert.True(t, pageInfo.HasNextPage)
+	})
+
+	t.Run("Should return the requested page of candles when last and before are provided", func(t *testing.T) {
+		last := int32(10)
+		before := entities.NewCursor(allCandles[100].PeriodStart.Format(time.RFC3339Nano)).Encode()
+		pagination, err := entities.NewCursorPagination(nil, nil, &last, &before)
+		require.NoError(t, err)
+		candles, pageInfo, err := candleStore.GetCandleDataForTimeSpan(context.Background(), candleId, &startTime, nil, pagination)
+		require.NoError(t, err)
+		require.Equal(t, 10, len(candles))
+		assert.Equal(t, allCandles[90], candles[0])
+		assert.Equal(t, allCandles[99], candles[9])
+		assert.Equal(t, pageInfo.StartCursor, entities.NewCursor(allCandles[90].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.Equal(t, pageInfo.EndCursor, entities.NewCursor(allCandles[99].PeriodStart.Format(time.RFC3339Nano)).Encode())
+		assert.True(t, pageInfo.HasPreviousPage)
+		assert.True(t, pageInfo.HasNextPage)
+
+	})
 }